icc complains conversion from 'long long' to 'float'
when -Wimplicit-const-int-float-conversion is used.

 ncx.c:2824:12: warning: implicit conversion from 'long long' to 'float' changes value from 9223372036854775807 to 9223372036854775808 [-Wimplicit-const-int-float-conversion]
         if (xx == LONGLONG_MAX)      *ip = LONGLONG_MAX;
                ~~ ^~~~~~~~~~~~
 ncx.c:64:22: note: expanded from macro 'LONGLONG_MAX'
 #define LONGLONG_MAX LONG_LONG_MAX
                      ^~~~~~~~~~~~~
 ncx.c:61:23: note: expanded from macro 'LONG_LONG_MAX'
 #define LONG_LONG_MAX LLONG_MAX
                       ^~~~~~~~~
 /opt/intel/oneapi/compiler/2023.1.0/linux/lib/clang/16/include/limits.h:103:20: note: expanded from macro 'LLONG_MAX'
 #define LLONG_MAX  __LONG_LONG_MAX__
                    ^~~~~~~~~~~~~~~~~
 <built-in>:52:27: note: expanded from here
 #define __LONG_LONG_MAX__ 9223372036854775807LL
                           ^~~~~~~~~~~~~~~~~~~~~

+++ src/utils/ncmpidump/ncmpidump.c
@@ -582,10 +582,10 @@ do_ncdump(const char *path, struct fspec* specp)
                                 var.fillval = NC_FILL_DOUBLE;
                                 break;
                             case NC_INT64:
-                                var.fillval = NC_FILL_INT64;
+                                var.fillval = (double)NC_FILL_INT64;
                                 break;
                             case NC_UINT64:
-                                var.fillval = NC_FILL_UINT64;
+                                var.fillval = (double)NC_FILL_UINT64;
                                 break;
                             default:
                                 break;
